refactor rotate into coordinate systems
