+++ geometric/coordinate_systems/rotate.py
@@ -10,14 +10,14 @@ References
 1. L.-P. Wang & C. C. Song, "Geometry optimization made simple using translation and rotation coordinates." J. Chem, Phys. 144, 214108.
 1. E. A. Coutsias, C. Seok, K. A. Dill. "Using quaternions to calculate RMSD.". J. Comput. Chem 2004.
 2. K. B. Petersen, M. S. Pedersen. "The Matrix Cookbook", 2012.
-3. G. H. Golub, V. Pereyra. "The differentiation of pseudo-inverses and 
+3. G. H. Golub, V. Pereyra. "The differentiation of pseudo-inverses and
 nonlinear least squares problems whose variables separate." Siam J. Numer. Anal., 1973.
 
 Copyright 2016-2020 Regents of the University of California and the Authors
 
 Authors: Lee-Ping Wang, Chenchen Song
 
-Contributors: 
+Contributors:
 
 Redistribution and use in source and binary forms, with or without modification,
 are permitted provided that the following conditions are met:
@@ -54,14 +54,12 @@ import numpy as np
 from geometric.molecule import *
 from geometric.nifty import invert_svd, logger
 
-"""
-"""
 
 def build_correlation(x, y):
     """
     Build the 3x3 correlation matrix given by the sum over all atoms k:
     xk_i * yk_j
-    
+
     Parameters
     ----------
     x : numpy.ndarray
@@ -85,7 +83,7 @@ def build_F(x, y):
     """
     Build the 4x4 F-matrix used in constructing the rotation quaternion
     given by Equation 10 of Reference 1
-    
+
     Parameters
     ----------
     x : numpy.ndarray
@@ -127,12 +125,12 @@ def al(p):
     Given a quaternion p, return the 4x4 matrix A_L(p)
     which when multiplied with a column vector q gives
     the quaternion product pq.
-    
+
     Parameters
     ----------
     p : numpy.ndarray
         4 elements, represents quaternion
-    
+
     Returns
     -------
     numpy.ndarray
@@ -146,18 +144,18 @@ def al(p):
                      [ p[1],  p[0], -p[3],  p[2]],
                      [ p[2],  p[3],  p[0], -p[1]],
                      [ p[3], -p[2],  p[1],  p[0]]])
-     
+
 def ar(p):
     """
     Given a quaternion p, return the 4x4 matrix A_R(p)
     which when multiplied with a column vector q gives
     the quaternion product qp.
-    
+
     Parameters
     ----------
     p : numpy.ndarray
         4 elements, represents quaternion
-    
+
     Returns
     -------
     numpy.ndarray
@@ -173,12 +171,12 @@ def conj(q):
     """
     Given a quaternion p, return its conjugate, simply the second
     through fourth elements changed in sign.
-    
+
     Parameters
     ----------
     q : numpy.ndarray
         4 elements, represents quaternion
-    
+
     Returns
     -------
     numpy.ndarray
@@ -196,12 +194,12 @@ def conj(q):
 def form_rot(q):
     """
     Given a quaternion p, form a rotation matrix from it.
-    
+
     Parameters
     ----------
     q : numpy.ndarray
         4 elements, represents quaternion
-    
+
     Returns
     -------
     numpy.array
@@ -212,9 +210,9 @@ def form_rot(q):
     return R4[1:, 1:]
 
 def sorted_eigh(mat, asc=False):
-    """ 
-    Return eigenvalues and eigenvectors of a symmetric matrix 
-    in descending order and associated eigenvectors. 
+    """
+    Return eigenvalues and eigenvectors of a symmetric matrix
+    in descending order and associated eigenvectors.
 
     This is just a convenience function to get eigenvectors
     in descending or ascending order as desired.
@@ -223,7 +221,7 @@ def sorted_eigh(mat, asc=False):
     if asc:
         idx = L.argsort()
     else:
-        idx = L.argsort()[::-1]   
+        idx = L.argsort()[::-1]
     L = L[idx]
     Q = Q[:,idx]
     return L, Q
@@ -232,7 +230,7 @@ def calc_rmsd(x, y):
     """
     Calculate the minimal RMSD between two structures x and y following
     the algorithm in Reference 1.
-    
+
     Parameters
     ----------
     x : numpy.ndarray
@@ -249,7 +247,7 @@ def calc_rmsd(x, y):
     y = y - np.mean(y,axis=0)
     N = x.shape[0]
     L, Q = sorted_eigh(build_F(x, y))
-    idx = L.argsort()[::-1]   
+    idx = L.argsort()[::-1]
     L = L[idx]
     Q = Q[:,idx]
 
@@ -259,7 +257,7 @@ def calc_rmsd(x, y):
 
 def is_linear(x, y):
     """
-    Returns True if molecule is linear 
+    Returns True if molecule is linear
     (largest eigenvalue almost equivalent to second largest)
     """
     x = x - np.mean(x,axis=0)
@@ -275,7 +273,7 @@ def get_quat(x, y, eig=False):
     """
     Calculate the quaternion that rotates x into maximal coincidence with y
     to minimize the RMSD, following the algorithm in Reference 1.
-    
+
     Parameters
     ----------
     x : numpy.ndarray
@@ -306,7 +304,7 @@ def get_rot(x, y):
     Calculate the rotation matrix that brings x into maximal coincidence with y
     to minimize the RMSD, following the algorithm in Reference 1.  Mainly
     used to check the correctness of the quaternion.
-    
+
     Parameters
     ----------
     x : numpy.ndarray
@@ -336,7 +334,7 @@ def get_R_der(x, y):
     """
     Calculate the derivatives of the correlation matrix with respect
     to the Cartesian coordinates.
-    
+
     Parameters
     ----------
     x : numpy.ndarray
@@ -347,7 +345,7 @@ def get_R_der(x, y):
     Returns
     -------
     numpy.ndarray
-        u, w, i, j : 
+        u, w, i, j :
         First two dimensions are (n_atoms, 3), the variables being differentiated
         Second two dimensions are (3, 3), the elements of the R-matrix derivatives with respect to atom u, dimension w
     """
@@ -380,7 +378,7 @@ def get_F_der(x, y):
     """
     Calculate the derivatives of the F-matrix with respect
     to the Cartesian coordinates.
-    
+
     Parameters
     ----------
     x : numpy.ndarray
@@ -391,7 +389,7 @@ def get_F_der(x, y):
     Returns
     -------
     numpy.ndarray
-        u, w, i, j : 
+        u, w, i, j :
         First two dimensions are (n_atoms, 3), the variables being differentiated
         Second two dimensions are (4, 4), the elements of the R-matrix derivatives with respect to atom u, dimension w
     """
@@ -445,7 +443,7 @@ def get_q_der(x, y, second=False, fdcheck=False, use_loops=False):
     """
     Calculate the derivatives of the quaternion with respect
     to the Cartesian coordinates.
-    
+
     Parameters
     ----------
     x : numpy.ndarray
@@ -462,11 +460,11 @@ def get_q_der(x, y, second=False, fdcheck=False, use_loops=False):
     Returns
     -------
     numpy.ndarray
-        u, w, i: 
+        u, w, i:
         First two dimensions are (n_atoms, 3), the variables being differentiated
         Third dimension is 4, the elements of the quaternion derivatives with respect to atom u, dimension w
     numpy.ndarray (if second=True)
-        u, w, a, b, i: 
+        u, w, a, b, i:
         First four dimensions are (n_atoms, 3, n_atoms, 3), the variables being differentiated
         Fifth dimension is 4, the elements of the quaternion second derivatives with respect to atom u, dimension w, atom a, dimension b
     """
@@ -515,7 +513,7 @@ def get_q_der(x, y, second=False, fdcheck=False, use_loops=False):
             dq2 = np.einsum('uwpr,abrs,s->uwabp', dinv, dF, q, optimize=True)
             dq2 += np.einsum('pr,abrs,uws->uwabp', Minv, dF, dq, optimize=True)
             # print(time.time()-t0)
-            
+
     if fdcheck:
         # If fdcheck = True, then return finite difference derivatives
         h = 1e-6
@@ -602,7 +600,7 @@ def get_expmap(x, y):
     """
     Calculate the exponential map that rotates x into maximal coincidence with y
     to minimize the RMSD.
-    
+
     Parameters
     ----------
     x : numpy.ndarray
@@ -623,7 +621,7 @@ def get_expmap(x, y):
 
 def get_expmap_der(x, y, second=False, fdcheck=False, use_loops=False):
     """
-    Given trial coordinates x and target coordinates y, 
+    Given trial coordinates x and target coordinates y,
     return the derivatives of the exponential map that brings
     x into maximal coincidence (minimum RMSD) with y, with
     respect to the coordinates of x.
@@ -644,11 +642,11 @@ def get_expmap_der(x, y, second=False, fdcheck=False, use_loops=False):
     Returns
     -------
     numpy.ndarray
-        u, w, i: 
+        u, w, i:
         First two dimensions are (n_atoms, 3), the variables being differentiated
         Third dimension is 3, the elements of the exponential map derivatives with respect to atom u, dimension w
     numpy.ndarray (if second=True)
-        u, w, a, b, i: 
+        u, w, a, b, i:
         First four dimensions are (n_atoms, 3, n_atoms, 3), the variables being differentiated
         Fifth dimension is 3, the elements of the exponential map second derivatives with respect to atom u, dimension w, atom a, dimension b
     """
@@ -715,7 +713,7 @@ def get_expmap_der(x, y, second=False, fdcheck=False, use_loops=False):
                     if maxerr > 1e-7:
                         logger.info("q %3i %3i : maxerr = %.3e %s\n" % (p, r, maxerr, 'X' if maxerr > 1e-5 else ''))
                     # logger.info("q %3i %3i : analytic %s numerical %s maxerr = %.3e %s" % (i, j, str(dvdq2[i, j]), str(FDiffV2), maxerr, 'X' if maxerr > 1e-4 else ''))
-                
+
     # Dimensionality: Number of atoms, number of dimensions (3), number of elements in q (4)
     if second:
         dqdx, dqdx2 = get_q_der(x, y, second=True)
@@ -799,7 +797,7 @@ def get_expmap_der(x, y, second=False, fdcheck=False, use_loops=False):
                             if maxerr > 1e-8:
                                 logger.info("atom %3i %s, %3i %s : maxerr = %.3e %s\n" % (u, 'xyz'[w], a, 'xyz'[b], maxerr, 'X' if maxerr > 1e-6 else ''))
             dvdx2 = FDiffV2
-            
+
     if second:
         return dvdx, dvdx2
     else:
